refactor(poller): move the invalid communication type error to the Dbal layer
The exception reports a value read from a database column, names the table in
its docblock and the lookup in its factory — knowledge the domain layer has no
business holding. Its neighbours in Domain/Exception stay agnostic of storage
even though the infrastructure is what throws them: AggregateNotFoundException
carries abstract criteria and a 'Resource not found.' message, and none of them
knows a database exists.

Only DbalPollerTransformer and its test referenced it, both already in the
infrastructure layer, so no domain class points at it any more.

Refs: MON-207613
